Breakout CA into a separate war
