+++ hopsworks-api/src/main/java/io/hops/hopsworks/rest/application/config/ApplicationConfig.java
@@ -12,7 +12,6 @@ public class ApplicationConfig extends ResourceConfig {
    */
   public ApplicationConfig() {
     register(io.hops.hopsworks.api.agent.AgentResource.class);
-    register(io.hops.hopsworks.api.agent.AgentService.class);
     register(io.hops.hopsworks.api.elastic.ElasticService.class);
     register(io.hops.hopsworks.api.exception.mapper.AccessControlExceptionMapper.class);
     register(io.hops.hopsworks.api.exception.mapper.AppExceptionMapper.class);
